# spec/openapi3.yml
openapi: 3.0.3
info:
  title: Unofficial and insufficient fincode API Spec
  version: 0.0.1
servers:
  - url: 'https://api.test.fincode.jp/v1'
security:
  - BearerAuth: []
paths:
  /customers:
    post:      
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                name:
                  type: string
                email:
                  type: string
                phone_cc:
                  type: string
                phone_no:
                  type: string
                addr_city:
                  type: string
                addr_country:
                  type: string
                addr_line_1:
                  type: string
                addr_line_2:
                  type: string
                addr_line_3:
                  type: string
                addr_post_code:
                  type: string
                addr_state:
                  type: string                                    
              required:
                - id
                - name
                - email                
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  email:
                    type: string
                  phone_cc:
                    type: string
                    nullable: true
                  phone_no:
                    type: string
                    nullable: true
                  addr_city:
                    type: string
                    nullable: true
                  addr_country:
                    type: string
                    nullable: true
                  addr_line_1:
                    type: string
                    nullable: true
                  addr_line_2:
                    type: string
                    nullable: true
                  addr_line_3:
                    type: string
                    nullable: true
                  addr_post_code:
                    type: string
                    nullable: true
                  addr_state:
                    type: string
                    nullable: true
                  card_registration:
                    type: string
                    nullable: true
                  derectdebit_registration:
                    type: string
                    nullable: true
                  created:
                    type: string
                    nullable: true
                    pattern: '^(\d{4})/(\d{2})/(\d{2}) (\d{2}):(\d{2}):(\d{2}).(\d{3})$'
                  updated:
                    type: string
                    nullable: true
                    pattern: '^(\d{4})/(\d{2})/(\d{2}) (\d{2}):(\d{2}):(\d{2}).(\d{3})$'
                required:
                  - id
                  - name
                  - email
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        error_code:
                          type: string
                        error_message:
                          type: string
                      required:
                        - error_code
                        - error_message
                required:
                  - errors
  /customers/{id}:
    delete:
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  delete_flag:
                    type: string
                required:
                  - id
                  - delete_flag
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        error_code:
                          type: string
                        error_message:
                          type: string
                      required:
                        - error_code
                        - error_message
                required:
                  - errors
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer